Renamings, extracted connectRouteHandlers to func
